move to MockUtils from ProductSyncMockUtils.
