+++ src/test/java/com/commercetools/sync/products/ProductSyncMockUtils.java
@@ -13,7 +13,6 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.sphere.sdk.categories.Category;
 import io.sphere.sdk.channels.Channel;
 import io.sphere.sdk.customergroups.CustomerGroup;
-import io.sphere.sdk.customers.Customer;
 import io.sphere.sdk.models.Asset;
 import io.sphere.sdk.models.LocalizedString;
 import io.sphere.sdk.models.Reference;
@@ -34,7 +33,6 @@ import io.sphere.sdk.states.State;
 import io.sphere.sdk.taxcategories.TaxCategory;
 import io.sphere.sdk.types.CustomFields;
 import io.sphere.sdk.types.Type;
-
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -505,19 +503,5 @@ public class ProductSyncMockUtils {
         when(customObjectService.fetchCachedCustomObjectId(any()))
             .thenReturn(CompletableFuture.completedFuture(Optional.of(id)));
         return customObjectService;
-    }
-
-    /**
-     * Creates a mock {@link Customer} with the supplied {@code id} and {@code key}.
-     *
-     * @param id  the id of the created mock {@link Customer}.
-     * @param key the key of the created mock {@link CustomerGroup}.
-     * @return a mock customerGroup with the supplied id and key.
-     */
-    public static Customer getMockCustomer(final String id, final String key) {
-        final Customer customer = mock(Customer.class);
-        when(customer.getId()).thenReturn(id);
-        when(customer.getKey()).thenReturn(key);
-        return customer;
     }
 }
